Change how non-bufferify functions are written
The bufferify version of a function will continue to deal with the len
and len_trim annotations.  The non-bufferify will now expect the
caller to deal with terminating NULLs.  The wrapper deals with name
mangling and char * to std::string conversion only.  Added tests for
calling the non-bufferified version of some routines.