Change the --sitedir option to --cmake
Create the cmake file instead of return a path to it.
This works better when installing shroud as a egg or pex file.
The --sitedir was returing a path to the file within a zip file.

Remove code to run uncrustify from the cmake file.
This is very Axom specific.