Rename Python temporary variables
Use pytmp_var for the object parsed and py_var for the NumPy array
object created from it.  This keeps py_var consistent with BuildValue
and return.

Make array variables as PyArrayObject instead of PyObject. This
removes the need for casts in PyArray_DATA but requires casting the
result of PyArray_NewLikeArray.