Change how namepaces are accessed
Instead of adding `namespace outer {` around the generated code, call
each function with explicit namespace, `outer:function()`.  For the C
interface, all of the functions are `extern "C"` so the namespace is
not needed. Likewise, Python has file static functions and one `extern
"C"` function.  The wrappers use the library to be wrapped but are not
considered part of the library so they should not pollute its
namespace with its own symbols.

namespaces are also required to be explict in types.  cxx_type and
c_to_cxx will require explicit namespaces.