Change how classes are foward declared.
When a class is declare without declarations, it is a forward declare.
The other fields such as options and format must be defined.  The
class can be redeclared latter with an additional declartions field.

This allows the type to be fully defined at the initial decl.  Before
a typedef was created with a dummy typemap.  Now that the typemap is
saved in the the AST, it must be the full typemap and not a typedef
typemap which will be replaced when the call was fully defined. Avoids
having to patch up the AST to remove to original typedef typemap.