Change how functions which return classes are wrapped
Add a pointer to a C_capsule_data_type argument.  It will be filled in
by the C wrapper which will then return the same value. (similar to
strcpy's destination argument) This allows it to be used in expression
to chain/nest function calls.

This also avoids returning a struct from C to Fortran which may
present some portability issues.