Create a capsule derived type for each class
XLF reported an error when F_capsule_data_type was being used in
multiple modules:
  Actual argument attributes do not match those specified by an
  accessible explicit interface.
Even if the derived types have the same members and same name, they are not
the same type if they are defined in different modules.

Now create a capsule derived type for each class.
The other alternative was to create a shared type in its own module.
However, this presents compile problems for the user since only one
instance of the capsule type can be compiled.