Add the external attribute
Uses the EXTERNAL keyword for function pointer argument.
This allows any function to be used as the dummy argument.

Add test for function pointers which uses function in a module.  These
function will have a implicit interface.  The C interface for the
function is bind(C) and uses an abstract interface which requires that
any function pointer dummy argument also have the bind(C) attribute.
EXTERNAL subverts this type checking by losing the interface.