Update wrapper for 'char *arg+intent(out)'
Update c_statement[intent_out_buf] to not create a local variable to
pass to the C function.  Instead pass the Fortran variable directly.
Add helper ShroudStrBlankFill which will blank fill the argument
before returning from the C wrapper.

Fortran must provide a CHARACTER argument which is at least as long as
the amount that the C function will write into.  This includes space
for the terminating NULL which will be converted into a blank for
Fortran.