Work on garbage collection for classes
Add a dtor member to classes which contain a struct with a
function pointer to a routine to release the array.
If dtor is NULL, do not release.
Removed explicit call to delete, since C structs are also
treated as classes (calls free).

Add additional error checking on memory allocations.