Add options.F_create_generic
User settable to control if a generic function is created. Useful for
templated functions when the template is part of the implementation
instead of the interface since all of the instantations will have the
same arguments.

Remove FunctionNode._CXX_return_templated - It prevented generics when
the return type was templated.  F_create_generic is more general.