Test fstatements in vectors.yaml
Return the number of items copied into the intent(out) argument.  An
array is passed in, then filled from the C routine.  Test two
variations.  One sets fstatements for both the C and Fortran wrapper.
The only only sets fstatements for the Fortran wrapper.

The previous commit contributed changes to allow this to work.