Move statements.declare into splicer
This allows a splicer to replace all statement code.  The function
arguments and result declaration are outside the splicer since they
should not be changed via statements but instead via decl and
attributes.

This required moving error checks for the dimension argument so that
vector function results get dimensions set to (:).

Added typemap.f_type_allocatable to get results declared as
CHARACTER(len=:) instead of CHARACTER(*).