Add FStmts.f_attribute
Allow additional Fortran attributes to be added to the declaration.
Since the declaration is outside the splicer, there is no other way
for the user to add attributes.  Needed when C_LOC is used since the
argument must have the TARGET attribute.

By assigning fmt.stmt0 where it is, some arguments such as implied
will not be assigned.  This keeps the JSON files comparing.

This change allows the test-fortran-desc to compile.