Move includes into module header file
Before each class was forward declared to avoid including the header.
However, it is not easy to forward declare templated classes such as
std::vector which must not only include the templated argument, which
will be in the YAML file, but also the defaulted template arguments
such as the Allocator.

Headers moved from the class implementation files.