Change helper x to be scope=utility
Then change to put the struct declared by the helper into the module
header file.  This is instead of duplication the struct in several
source files.  Prefix the struct with {C_prefix} to allow multiple
packages to co-exist.  (There actually may not be a problem, but
having unique names may avoid confusion.)