Use get_indirect in default search
For all cases it does not make a difference.

Was py_native_in, now py_native_*_in but uses py_default
Format fields are used to make them usable by both pointer and
nonpointer. {deref}, {addr}

But it will be useful with arrays. py_native-[]-in which
must copy memory.