Fix more issues with pointer struct members
Treat pointers as an array even if no dimension attribute.  It is
still possible to assign a list or a NumPy array.

Revert previous change which added readonly attribute for
'const double *'.  But constness is cast away when creating the
NumPy array in the getter.

Remove py_statements py_descr_native_*_scalar since the existing
numpy and list variants work.