Improve dealing with const in capsule
Change the way a capsule holds a void * to now be a union:
    union {
        void *base; /* address of C++ memory */
        const void *cbase;
    } addr;

This allows lots of casts to be removed from the wrappers by correctly
assigning to the const or non-const member.  The correct member is
defined by fmt.capsule_addr.
Removed lots of expliciting casting to void * before assiging to capsule.
Removed fmt.cxx_cast_to_void_ptr which tried to deal with C++ const casting:
-    SHC_rv->addr = static_cast<void *>(const_cast<classes::Class1 *>(SHCXX_rv));
+    SHC_rv->addr.cbase = SHCXX_rv;
Where base is 'const void *'.

Partly this is fixed by realizing that C++ can assign to void *
without a cast (like C).  But the const_cast would still be required
to avoid the error:
  error: invalid conversion from 'const void*' to 'void*' [-fpermissive]
This would still be a problem in C, since gcc reports the warning:
  warning: assignment discards 'const' qualifier from pointer target type

Fix assigning to fmt.c_const (and now fmt.capsule_addr) that confused
a const method with a const function result.

The Lua code was changed. But nothing has been compiled so can't be sure
about it.