Remove call to typemap.compute_return_prefix
Instead pass variable as declared.  Rely on PyStmts.arg_call
to do any non-default dereference/address.
This removes a use of cxx_local_var and c_local_var.

Add spointer when looking for struct statements.  Note it is always a
pointer in the wrapper since the PyObject contains a pointer to a
struct.