Add PyStmts.fmtdict
Used to override definition of ctor_expr when needed.  The alternative
was to add PyStmts.ctor_expr but adding fmtdict will make it possible
to override other variables as needed without adding more fields to
PyStmts.

Remove resetting c_deref which changed the value of ctor_expr.  Needed
an explicit way to 'fix' ctor_expr.

c_deref was being reset because c_local_var defaulted to "scalar", but
that is not always true now since a declaration is added with the same
type as the argument.

Only used with py_native_*_out to keep generated code the same.
Only JSON files changed.