Add PY_member_data to struct-as-class
Holds PyCapsule object which owns the memory.  However this object is
not returned by the getter. Instead a new object will be created.
PY_member_object holds the cached object returned to the user which
must be something like a NumPy array so that when the object changes
the C memory changes as well.

Used with py_descr_char_* and py_descr_char_**_list for now.