Mangle name of Fortran helper capsule_dtor
If multiple Shroud wrapped modules are used in a file, they will each
contain an interface for the Fortran helper.
Ideally the interface would be PRIVATE.
However, gfortran does not allow PRIVATE and BIND(C) to be mixed since it
sees them as contradictory.
Instead, mangle the helper name with C_prefix to make them unique.