Set name of variant statements
When the variant is something like "scalar/*", i.e. takes a scalar or
a pointer, the statement name is added to the source when the debug
option is set.

Howewever, when preprocessing the generated source the /* looks like a
C comment and cause problem even when part of a Fortran comment.
Rename the generated statements dict to use the specific name,
"scalar" or "*" to avoid the problem.

This may make it more difficult to find the code in fc_statements for
an exact match since the name is now generated. i.e. grep will not
work.