Create all Fortran wrappers before bind(C) interfaces
The Fortran wrapper may change the name of a bind(C) interface when
the Fortran wrapper is not needed so the bind(C) can be called
directly.

Before the wrapping was done per function but this will cause a
problem when a generated C wrapper is processed before it's
corresponding Fortran wrapper.

Changes the order of wrapping in the log files.