Set metaattrs["deref"] for functions
This work was done by check_return_pointer after all functions were
generated.  Now it is set in check_deref_attr_func when attributes are
initially checked before any additional FunctionNodes are created.

The most tweeking is done when the result is converted into an
argument since the deref must move from the function to the additional
argument.

Some additional bufferify functions are created which will pass a
context argument. This maps better to the CFI_cdesc_t which is used in
Fortran 2018.