Add test python-only
tutorial.yaml was creating fortran and C wrappers because of a recent
change which promoted wrap_fortran up to the top level.
has_default_args was creating functions for Fortran and C when the
wrapper was not requested.

Add a regression test to process tutorial with only wrap_python=True.

#226