Only set arg.stmts_suffix if needed
Before, if any argument needed it to be set, all arguments for a
function were set.  This caused scalars to get set as buf when it was
unneeded.  This will compute the wrong statements name.  They were
defaulting to c_default so no generated code changed.
Ex. c_native_*_inout_buf is now c_native_*_inout.  No buffer argument
is passed or needed.