Improve how CFI CHARACTER functions are handled
Instead of adding an extra argument for the function result, do all
the work via statements. Makes this more data-driven instead of
relying on code. The downside is that the C wrapper will be a 'char *'
function where the result will be thrown away since all the work is
done to the CFI_cdesc_t argument.

Add c_stmts.f_c_arg_names - list of arguments of bind(C) interface.