Set fmt.c_array_size with new buf statements
To set fmt.c_array_size it was looking for attrs["context"] which is
not used in the new statements. Instead {temp0} will be
used. Implemented as a kludge since it assume temp0 only.  Need some
semantic way to name variable in statements maybe.

Change an exception in wrapf.py into a message in the log file as well
as a line in the generated code.  This will help ensure the file is
created even if there is an error.