Move C_memory_dtor_function to file C_impl_utility
Before this change it was written into C_impl_filename. It contains
code to release memory from shadow classes.  It has code from all
wrapped classes, as well as the STL. It makes more sense to put it
into a shared file instead of the file with wrapped library functions.