Replace buf_extra["shadow"] with mixin
Makes fc_statements more explicit and more data driven.

buf_extra added a result argument at the end which was used to return
non-native types, like classes, as shadow types. This is now done
explicitly via the c_function_shadow statement group. The return
argument is now first (but after the self argument).

C wrappers no longer return a pointer to the shadow struct and instead
are ``void`` functions.  The shadow struct must continue to be passed
in as an argument which will be filled in by the C wrapper.  The main
reason is that the Fortran wrapper had to have a local type(C_PTR)
variable to save the result but then never did anything with it since
all the work was done with the shadow type.