Remove setting fmt fields from attributes
This has been converted to Stmts.ntemps to generate temporary variable
names. This is mainly removing unusable names. For example:
 char *getname() +len(30)
Sets fmt.c_var_len to 30 which is an illegal variable name.  There are
a few tests which changed the names via +len(AAlen) on a variable.
This is no longer supported.