Change how +cdesc attribute is handled
It is now processed by buf_to_arg and will set metaattrs["api"] to
"cdesc" for the argument. A bufferify function will be created.
Before, the unbufferify function had the +cdesc applied.  Now, the
function is wrapped directly allowing the C API to call the function.

"cdesc" is now a value for api.  It will eventually be used in place
of "buf" in statement groups which use c_var_desc.  "buf" will apply
to groups which add metadata from Fortran intrinsics such as LEN or
SIZE.  Removed code in wrapf and wrapc which used both "buf" and
"cdesc" in the statement group lookup.

Currently only the test case in cdesc.yaml is handled with api=cdesc
which applies to arguments. Eventually needs to be set implicitly for
wrappers which use c_var_desc. i.e. intent(out)

Add FunctionNode.splicer_group which will be set when the bufferify
function is created. Before the metaattrs["api"] of the function was
checked, but it will not be set now if only an argument is
bufferified.  So splicer 'c_buf' will be used. This reverts to
pervious behavior from March 17 and will ignore the splicer for the
non-bufferified version.