Set f_intent for bind(C) interfaces
This fixes a path through the code where fmt.f_intent was not set
resulting in Fortran code with "intent()" (which does not compile).
Moved setting of f_intent from set_fmt_fields to set_fmt_fields_iface.

Added a test case which missed this path. Create a generic with a
struct-as-class argument. This was derived from a Basis test case.

  - decl: long UpdateStructAsClass(StructAsClass *arg, long inew)
    fortran_generic:
    - decl: (int inew)
      function_suffix: _int
    - decl: (long inew)
      function_suffix: _long

A bind(C) interface is created for both generic wrappers (int and
long).  But it also creates a wrapper for the original function. This
one was missing the fmt.f_intent since it is not called by a Fortran
wrapper.  It probably doesn't need to exist.  The generic feature is
done by the C wrapper and automatic type promotion instead of the
other examples which call the INT intrinsic directly to convert in the
Fortran wrapper.