Remove .SECONDARY target from run/Makefile
This was required to preserve the directory used to compile a test.
Otherwise, an error was reported like:

rm build/temp.linux-x86_64-3.7/run/./tutorial/..
make: unlink: build/temp.linux-x86_64-3.7/run/./tutorial/..: Is a directory

Fixed by added directory creation on the rule to run the test.
Before it was only on the rule to compile the test which caused it to be
a secondary target.
Also converted to order-only dependency.
Requires gmake 3.80