Do not add hidden arguments to Fortran or C wrapper
Use statements to create a local variable for the hidden argument in
the C wrapper and pass to the library routine.  The native C api will
still have the hidden arguments since its main purpose is to call C++
routines from C and the C user will need the value.  Fortran uses the
hidden argument in ALLOCATE or C_F_POINTER.