Update fc_statements c_function_char_scalar
pgi and cray have issues with a bind(C) function which returns a
CHARACTER(len=1,kind=C_CHAR).  It must not be considered a scalar
result. Update statements to pass result as an argument to the C
wrapper which then assigns to *SH_cxx_rv to return a single char.

Set need_wrapper when CStmts.call is set.