Return Fortran pointer from some getter functions
If the class field is a pointer with a +dimension attribute, the
getter will return a Fortran pointer.

When creating the getter function, copy metaattrs from the member
field to the generated function. This will add +dimension to the
getter.

Check meta["dimension"] instead of attrs["dimension"] since the getter
does not copy user provided attributes, only shroud generated meta
attributes.  attrs["dimension"] is a string, meta["dimension"] is a
parse tree of the string.

Added statement c_getter_native_*_cdesc. Similar to statements for a
function but uses the class member directly instead of calling a
method.