Create getter and setter for pointer members of a struct
These are similar to class getter/setter.  Allow the user to get a
pointer from a struct provided that the dimension attribute is set.

Create LibraryNode.class_map.  The metaattr["struct"] is the name of
struct/class. This name can be looked up in class_map to find the AST.
Setting the metaattr to the AST node directly creates too much
recursion when printing the json tree.

Some tests changed by adding the process_class after creating a
default constructor for Python for a struct. The constructor did not
have a _function_index before.

Note that the C++ file for struct-cxx generates some warning errors
about unused variables.