Write wrappers for typedefs in C and Fortran
Create typedefs for C and Fortran wrappers which mirror the C++
typedef.
  typedef int TypeID
C:
  typedef int CLI_TypeID
Fortran:
  integer, parameter :: type_id = C_INT

Before the 'int' was used in the wrappers without any trace of the
typedef.

Typedefs in a templated class are instantiated.

Do not deepcopy Typemap. Only one true copy of each Typedef is created
and registered. This makes it possible to update templated typedefs.