Reorder assignments in fill_class/struct_typemap
Assign fields which are derived from other fields after updating
Typemap from fields dictionary.  This allows the typemap section of
the YAML file to specify a subset of fields and allow the rest to be
computed.  Many of the derived fields are for the convience of shroud
when generating code and would be redundant for the user to set.

Set Typemap.f_module_name for structs, similar to how it is set for
classes.  This changes the reference files.