Set user_fields for ast.TypedefNode
Record the user provided typemap fields so they may be reapplied if
needed.

Pass fields as an explicit argument instead of bundeling into *kwargs.
This removes some ambiguity when fields argument is passed through a
function via kwargs and makes grep easier. On the other hand, it
clutters things up a bit with extra arguments.