Control when to set default Typedefs fields
If the Typemap field is set via the fields dictionary, then do not
compute a default value. Typically only f_derived_type is explicitly
set. Then f_class, f_type, f_c_type, f_module and f_c_module are
computed.  This is for class and struct Typemaps.

Set default argument value for fields to {} instead of None to allow
it to always be treated as a dictionary (IN and UPDATE operations)