Do not create a bufferify wrapper for all function pointers results
The Fortran wrapper will deal with the type(C_PTR) to create a Fortran
pointer. char * still needs bufferify to deal with the length.

This will remove bufferify functions which are identical to the
non-bufferify function.

The IBM compiler complained about the non-bufferify and bufferify
interface both using bind(C) to the same C function.