Add statements.notimplemented field
If set to True, the generated wrapper will be surrounded in #if
0/#endif to show the generated wrapper. It's possible the wrapper will
not compile.

This came up with:
  string** strs +intent(out)+dimension(name_len)+deref(allocatable)

The c_default statements were incorrect since it tries to map
std::string** to char **. However, more work is needed for the C api.
By adding notimplemented to the c_out_string_**_allocatable
statements, the work needed to create a proper C wrapper can be
deferred.  Only the bufferify function was needed in this case.