Prefix c_statements call, ret, return_type
Rename ret as c_return.  It was shorted before because dict(return=x)
does not work since return is a reserved work.

Prefix the C groups variables with c_ so the Fortran and C groups can
be merged together. This allows a single group to describe the
Fortran-interface-C flow for a variable.