Create exact matches for argument c_statements in wrapf.wrap_function_impl
Creates additional groups so that the previously matched group is now
an exact match.  This will ensure that shroud knows how to wrap an
argument without accidently defaulting to an incorrect match.

For example:
-    ! Requested: c_out_native_*
-    ! Match:     c_default
+    ! Exact:     c_out_native_*

Several degenerated to using f_default. This only requires that the
group exists and f_default is automatically its parent.

Temporary checks in place marked by TTT comment.