fc_stmts names must match exactly
Changed the name to be looked up in the fc_dict dictionary instead of
cf_tree. If the name is not found, raise an exception.

In the past when a new statement group was added with a specific name,
it would often cause the tree walk for another argument to end up in
the wrong place since new branches had been added.  Now the statement
name must match exactly so there are no surprising changes.  It also
lets the user know when the argument type has not been wrapped before
since the statement will not exist instead of guessing incorrectly and
potentially creating bad code.

The statement tree is still created since the Python wrappers use it.