Merge group fc_out_string_**_cdesc_allocatable
By merging the two very specific parts, a fuller picture of how the
Fortran and C work together can be seen.

    dict(
        name="fc_out_string_**_cdesc_allocatable",
        alias=[
            "f_out_string_**_cdesc_allocatable",
            "c_out_string_**_cdesc_allocatable",
        ],

The c_helper lines had to be concatenated which changed the
fmt.hnamefunc0 value.

The aliases are needed since the wrapf and wrapc still look for them
but eventually, it should only look for the fc statement.