Use process_mixin to create groups
First base and mixin groups are processed. This allows a mixin to use
a mixin.  Before a mixin did not recurse.  Alias are saved then
processed after all of the permuted groups are created.  Default
values are applied first, then data from the group is updated in order
to make sure the defaults do not overwrite the values from statements.

The none/statements changed since some groups where changed because
their bases no longer exists since it is now a fc_ group.