Compute some statement name during generate
and set in declast.Declaration.fstmt.  Compare it against the current
scheme to make sure it is the same.

Define group f_function_char_*_cfi_arg which converts a function
result into an argument. To replace generating a new function in
arg_to_cfi.

Allow stmt.f_result to be *subroutine* to convert a function
into a subroutine.