Add fmt.f_c_suffix
Used in fmt.C_name_template and fmt.F_C_name_template to allow the C
function to have a mangling than the Fortran wrapper.  Used to be
appended to fmt.function_suffix but it makes more sense to be its own
field.  This will eventually remove the need for an extra Fortran
wrapper layer in gen_to_cfi and gen_to_buffer.