Reverting computing ast.fstmt
ast.fstmt was set in check_deref_attr_func by looking at
fmt.F_string_resuilt_as_arg and setting metaattr["deref"] = arg.
However, that is too soon.  It makes it difficult to generate a C
wrapper and a Fortran wrapper because deref will be wrong for the C
wrapper.