Remove alias from intent=mixin group
Ideally, a mixin group would only define a few fields since it will
never be used directly in code generation. No need to set defaults.
But first, a mixin can not have alias which may be used in code
generation.

Allow a 'shared' intent to make it easier to use alternate names for now.
        name="f_shared_function_string_scalar",
        alias=[
            "f_function_string_scalar/*/&",
            ... a long list of names with alternates

Used a check in process_mixin to find them.