Always use format name for helper in statements
-            "ShroudStrCopy({c_var}, {c_var_len},"
+            "{chelper_ShroudStrCopy}({c_var}, {c_var_len},"

Separates the helper name from the helper's function name allowing
them to be different by adding a layer of indirection. And allowing
the function name to be changed without changing the statements.

This also cleaned up the statements by only listing the helpers which
are actually used by a group.