Add error.cursor to record current position
This allows better error messages to be created.  In particular,
util.wformat replaces the SystemExit with a better error message with
context and returns a string which will be inserted into the generated
code (but will not compiler), then can continue processing.

Also replace a RuntimeError in Wrapf.wrap_class with a better error
message, set the format string to a dummy value, then continue
processing.