Add using statements for structs
When wrapping a C++ struct, use the library's struct type for the
function wrappers instead of C wrapped struct. The wrappers are
compiled with C++ so namespaces are not a problem. The types should be
identical, assuming it is described completely in the YAML file, which
is required for bind(C) to work.