Rename fmt fields for arguments
Renamed some format fields to allow more control of argument names in
wrappers.  The C wrapper continues to use *c_var* and *cxx_var*.  The
Fortran wrapper continues to use *f_var*, but if a different argument
is needed to be passed to the C wrapper it is now *fc_var* instead of
*c_var*.  The interface uses *i_var* instead of reusing *c_var*.

i_var is used in statements iarg_decl, i_arg_names and i_result_decl.