Add FunctionNode._generated_path
This list is computed as functions are created.  It replaces the
C_generated_path which was not always correct when CXX_PTR was not
updated consistently.

Functions which are created directly, not cloned from a YAML function,
such as getter/setter, fortran_generic and struct_as_class_ctor are
added to the list.