Redefine meaning of wrap.f and wrap.c
WrapC.wrap_function will get called once if wrap.c is True and again
if wrap.f is True.  i.e. wrap.f now controls when a C wrapper is
created if it is needed by the Fortran wrapper.

Creates a fmtc for C wrappers created by option.wrap_c and fmtf for
option.wrap_f.

Create a FunctionNode.C_signature to describe which statements have
been used to create a wrapper. This way if the wrap.fortran version of
the C wrapper is identical, then it will not be written again.

Do not add deref in lookup_c_function since (currently) deref only
describe how the pointer will be dereferenced in Fortran.
Moved c_function_void_* as an alias of f_function_void_* so two
identical functions will not be created.

This is progress on the ability to create the C wrappers intentionally
and not just as a side-effect of the Fortran wrappers.

Change Statement comments in wrappers depending on language wrapper
being created.
-// Statement: f_subroutine
+// Statement: c_subroutine