Factor fstatements into signature
If a function is generated with input from fstatements, then the C
wrapper for wrap.c and wrap.f are assumed to be different and a
function will be generated for each language wrapper since the
signature will not match.

The error.yaml test demonstrates by generating additional wrapper for
function BadStatements.