Distinguish between two stmt groups
f_mixin_function_ptr and f_mixin_funtion_c-ptr create identical C
wrappers but are different in the Fortran wrapper.  Add c_need_wrapper
to force a C wrapper to be created so that two fortran interfaces do
not bind the same C function since compilers do not like that.  The
signatures need to be different.

For examples, pointers-cxx creates two identical C wrappers for
returnIntPtrToScalar. One for C and one for Fortran with a _bufferify
suffix.