Process meta attributes from YAML for function arguments
The deref attribute is not used for C wrapper, only the meta attribute.

    decl: void func(std::vector<int> &arg+intent(out)+deref(allocatable))
    bind:
      c:
        decl: (arg+deref(malloc))

The fortran wrapper uses allocatable and the C wrapper uses malloc.
Note that the fortran wrapper also creates a C wrapper which uses
allocatable. This wrapper will set the api to cdesc.

Create statement groups for some vector arguments.