Move attr checks on function pointer arguments
Only intent and value are checked.
Done as part of FillMetaShare, removed from generate.py.
attrs are updated directly unlike regular argument attributes
which are updated as a meta attribute.
The intent attribute is read while creating abstract interfaces
and value is used as part of param.bind_c() calls.