Set more attrs for getter/setter
intent, api, deref

Before the meta attribute was set directly.  Now, the meta attributes
are computed after the function is generated so only set the
attributes.

Changes to comments in generated code and JSON files.

Had to add a kludge to get cxxlibrary.yaml get_nested_child test to
compile.  A pointer to a pointer to a struct. The tests are commented
out so it was not working before.  But with new checks it was
complaining about added the deref attribute. If was forced in before
by setting metaattrs directly.