Change attributes to be dict instead of defaultdict
By using a dict, the user supplied values in the YAML file can be
preserved. The default dict created entries when queried making it
impossible to know if the user set it or it was a side effect of the
query.

Remove some old use of attr["allocatable"], now deref=allocatable.