Allow format field with typedef
Test with F_name_typedef and C_name_typedef.  Allow typedefs to be
renamed individually for each wrapper language.  This removes the need
for the field entry since all of the fields are now computed based on
F_name_typedefs and the base type of the typedef.

- decl: typedef int32_t IndexType2
  format:
    F_name_typedef: LOCAL_Index_Type
    C_name_typedef: LOCAL_IndexType
  #fields:  No longer needed

Rework add_typedef to use argument names (format, options) instead of
**kwargs.  It's a better practice. It was convenient to convert now since
format is now being passed down.  Eventually use with all of
add_declaration.