Progress on enum
Use the c_type for enums for the C wrappers.  Use the ci_type for the
bufferify wrappers.  Fortran treats enums as integers the bufferify
will cast to the correct enum. This also uses the ci_to_cxx as needed.

Lua and Python also use the ci_type since lua_pushinteger
andParseTupleAndKeyword require integers.

Set an enum's Typemap.c_type to "enum %s" instead of 'int'.
The C wrappers deal with enums, Fortran/bufferify with integers.

Remove the c_to_cxx and cxx_to_c for typedef Typemaps.  When a typedef
is an alias (typedef int foo), no cast is needed.  Instead the problem
this was fixing is solved by the converters in the enum Typemap.