Move fc_statements into a JSON file
Read the value of fc_statements from a JSON file.  This will make it
easier for users to change this one file if necessary instead of
editing statements.py. Emphasises that the process is data-driven.

Originally the intent was to make this a YAML file, since the input is
YAML too.  I find YAML easier to read since it has less punctuation.
However, it is also much slower.

Sample timings of the test suite
Original Python
19.532u 6.300s 0:29.84 86.5%	0+0k 20584+38624io 28pf+0w
YAML
35.757u 5.735s 0:45.30 91.5%	0+0k 11464+38624io 16pf+0w
JSON
19.495u 6.331s 0:34.55 74.7%	0+0k 16104+38624io 0pf+0w