Change pattern used to look up arguments in statements
Use a modified abstract declarator which uses the typemap.sgroup
instead of the type.

- c_in_vector_&_targ_native_scalar
+ c_in_vector<native>&

This uses a shorter name which maps more directly to the C++
declaration.  The original reason for separating the scalar/*/& from
the type no longer exists. (It was using the underscore delimited name
as a path to try to match to a default statement. The name must now
match exactly.)

Every where that outputs the satement name changes: debug comments in
wrappers and JSON files.