Use cxx_T1 in vector statements
In order to remove convert_vector_typemap, the statements must contain
all the info about the vector template argument.  This works in the C
wrapper, but Fortran still makes the assumption that the typemap has
been changed to use the template argument (since there is only one
argument supported currently).  i.e. vector<int> will add INTEGER
argument to the wrappers.  That info is still missing in the fmt
dictionary.