Define fmt.f_module_name
Taken from typemap.f_module_name if not None.  It defaults to
'-f_module_name'. This values is checked in update_f_module which will
skip it if it is the default value.  This will allow to be used more
generallay in statements f_module.

        "f_module": {
            "{f_module_name}": [
                "{f_kind}"
            ]
        },

Is a no-op when f_module_name is "-f_module_name-".

Changes JSON files by adding fmt.f_module_name.