Test USE of function pointer result
Create a function pointer which returns a double while the other args
are int. This avoids accidently getting the correct USE for C_INT for
the result when it was added by the arguments. Make sure the result is
adding it's own USE C_DOUBLE.

Derived from an example.yaml test which changed with the previous
commit.