Renamed format fields *hnamefunc* for descr
These fields were added from the statement field **getter_helper** and
**setter_helper**, which are a list of names. A format field was added
for each name with a 0-based suffix corresponding to the position in
the list.  Now, the format fields have the prefix of *c_helper_*
followed by the helpers name. For example, *c_helper_copy_array*.
This makes it easier to match the corresponding helper and will help
when using statement mixin groups since the order of names will no
longer matter.  This is similar to the **c_helper** entry.

The helper can also contain an alias, ``helper:alias``. This is needed
when the helper contains a format field such as ``{cxx_type}``.

Add better error reporting to wrap_class_variable.