Do not use option C_shadow_result with Fortran
It is only used by the C wrappers intended to be called by a user.
The C wrappers created to be called by the Fortran wrappers will never
return a value. The capsule is passed in as a function argument and
filled in. But the C wrapper will always be a void function.

Note that this doubles some wrappers. A User callable capptr version
and a Fortran callable capsule version.

This is to reduce the number of permutations that will be added when
+owner(shared) is implemented. Also, the returned pointer to the
capsule was never used by the Fortran wrapper.