Set meta["abstract"]
The abstract type from the declaration, ex. 'native' or 'native*'.
Before it was computed right before it was used to look up the
statements.  By storing it in meta attributes, it's easier to change
it if necessary. For example, with std::shared_ptr. This changes all
of the JSON files which dump the meta attributes.