prepare for shared_ptr functions
Replace the "base" entry in f_function_shadow_capsule_shared and
c_function_shadow_capptr_shared with the list of mixins used by the
base group. This will make it easer to adjust the groups by changing a
mixin to deal with std::shared_ptr.
(These groups are currently unused).