Fill out group c_function_string
Used with function which return a std::string by value.  The wrapper
must allocate a local variable and returns a pointer to the c_str in
the instance. The wrapper uses a capsule to hold the std::string which
the user must delete.

This removes the C_error_pattern in the YAML since it is part of the
group now.