Add group f_mixin_as-intent-out
Remove setting the intent to OUT for function results in fcfmt.py.
Instead use this group to explicitly set the intent to OUT when
converting a function result into a subroutine argument.
"Explicit is better than implicit."

Note that f_intent_attr is used with both Fortran and interface
arguments.